# v.1.0.7: Se migra de la versión netcoreapp3 a net5 y se actualizan librerías.

+++ Kitpymes.Core.EntityFramework/Extensions/EntityFrameworkServiceCollectionExtensions.cs
@@ -127,6 +127,11 @@ namespace Kitpymes.Core.EntityFramework
                 .AddDbContextPool<TDbContext>(settings.DbContextOptionsBuilder)
                 .ToService<TDbContext>();
 
+            if (context is null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
             if (settings.IsEnsuredDeletedEnabled == true)
             {
                 context.Database.EnsureDeleted();
